2018-08-28 - added: Support XGD3 dumping on kreon drive - fixed: Rereading of SecuROM/Libcrypt - deleted: some warning

+++ DiscImageCreator/_linux/defineForLinux.cpp
@@ -325,12 +325,12 @@ int WideCharToMultiByte(
 
 #define NO_ERROR 0L
 
-void SetLastError(unsigned long errcode)
+void SetLastError(int errcode)
 {
 	errno = errcode;
 }
 
-unsigned long GetLastError(void)
+int GetLastError(void)
 {
 	return errno;
 }
